Solved y2022 day 10
